import RxJS operators from the main RxJS path
As of [RxJS
7.2.0](https://github.com/ReactiveX/rxjs/blob/master/CHANGELOG.md#720-2021-07-05),
RxJS operators are now all exported from the top-level `"rxjs"` path.

The changelog even indicates that the previous behavior of importing
from `"rxjs/operators"` will soon be deprecated.

Because soon is better than never and because it wasn't that hard to do
(with the power or Vim at least!) I decided to do a quick commit/PR on
this.

Note that sadly this commit might bring a lot of conflicts when it will
be merged. Hopefully, they all should be pretty easy to merge.